changement au niveau d'entitie

+++ src/main/java/com/example/myproject/entities/Event.java
@@ -34,11 +34,12 @@ public class Event implements Serializable {
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue (strategy = GenerationType.IDENTITY)
-	private Long idEvent;
+	private int idEvent;
 	private String name;
 	private LocalDate date;
 	private String lieu;
 	private String description;
+
 
     @ManyToOne
 	@JsonIgnore
